+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- Updated template folder to latest version of geostreaming explore and download page
+
 ## 3.8.0
 - See `packages/core/CHANGELOG.md` and `packages/geostreaming/CHANGELOG.md`